requirements.txt
colorama==0.4.6
python-dotenv==1.0.0
openai==0.27.6
tiktoken==0.5.1